Collapse of improbability premise I: file clearance.
The speed of light has been claimed as a condition of life itself.
https://www.discovery.org/m/securepdfs/2018/12/List-of-Fine-Tuning-Parameters-Jay-Richards.pdf
https://en.wikipedia.org/wiki/Fine-tuned_universe

This undermines the premise of the improbability of the cosmic fit.
From the fact of life, the speed of light would come as no surprise.
Present interstellar distances likewise: in an expanding universe,
they would eventually be reached.  Since those are the only conditions
of the cosmic fit, it too is no surprise.